requirements.txt
yt-dlp
discord
python-dotenv
